Standardise preview warning box when using ?action=edit
If editing without JS, and clicking preview on action=edit
you will see a preview with a note at the top of the page.
McrUndoAction.php uses a close copy of the same code and will
be amended accordingly.

Currently this is using a class `.previewnote` and not the
standardised and semantic "warningbox".

The messages of the preview will now be warningbox's
- see screenshots for end result:

* https://phabricator.wikimedia.org/F30284946
* https://phabricator.wikimedia.org/F30284947

Bug: T232414
Change-Id: Ib678f03e908fa9dcd5af7632941475203478abf7