docs: Disable GENERATE_TESTLIST and GENERATE_BUGLIST
These are to create pages with an overview of `@bug` and `@test`
annotations of which there are none in MediaWiki core in doc block
comments, like `/**`. As far as I know, there is also no practice
or plans to start using these or to use Doxygen as a way to
organise such work (we use Phabricator instead to report things
that are broken and/or should have tests).

Bug: T232104
Change-Id: I1056f22c38b11157ac7b5a23e9838978e8901171