docs: Ignore extensions/ and skins/ in mwdocgen.php by default
The default was previously to index all of core and extensions,
which is likely not the intent of someone running this for the
first time (which can then easily take over half an hour).

Even in CI, we never do this, and we never have afaik.

Generally, local use is to replicate CI (all of core), or a subset
of core by passing e.g. '--file includes/libs' or some such.

Indexing an extension is still possible, same way as before,
by passing '--file extensions/FooBar'. And, if one really wants
to index all of core and all extensions, the new '--extensions'
option allows one to opt back in to the old behaviour.

Also:
* Exclude all of 'resources/'. There are no PHP or Markdown files
  here that need to be indexed.

Bug: T232104
Change-Id: I55617776ac86f3fdcb6c1ba1eebe892edcb5fafb