wdio-mediawiki: extract bot() function
This functionality was previously copy+pasted for most other methods,
and is also useful externally, to make API requests that are not covered
by other methods in Api.js.

createAccount() needs a different type of token, so for now we leave it
alone. (Perhaps the token type could become a parameter of bot() later?)

Change-Id: Ica24213a1f0d651bfc3fbffe24b63251a4841658