Remove Revision::getRevisionText from migrateArchiveText
The script is used to migrate archive text from pre-1.5 schema.
It's safe it just use the `ar_text` field directly instead of going
to Revision::getRevisionText since migrating this field is the whole
proing of the script. Per execution it checks that the field exists
in the schema, the SQL query reads the field.

We don't need to do `expandBlob` either, cause we already know that
the data is not in external store, thus Revision::decompressData
is enough.

Bug: T198343
Change-Id: Ic838bdf680d73522dd508dd86056eb945535b368