selenium: Improve README docs
Simplify the README to only mention the bare minimum about
how to use WDIO itself. The rest is deferred to the various wiki
pages, which are more up to date.

Also removes selenium.sh file, which is no longer used as of 1955a8aa564,
I forgot to re-remove it during one of the rebases.
It is now supported to pass arguments to 'npm run selenium' as well.
Previously, arguments were only supported for 'npm run selenium-test',
which required manual starting of a chromedriver.

Bug: T214708
Change-Id: I5951af9146ac5fdce8d6506d703620049d6efe74