special.version.css: Replace problematic BIDI attribute with dir
This is an alternative for Ifb739d8, fixing the same problem.

The CSS attribute was introduced in 2014 via Iff39c73. In theory, it
should work, but appears to trigger bugs in certain browsers. Relying
on the dir="…" attribute is a solution that is already used all over our
code bases. It appears to be much more stable, resulting in the same
output.

Bug: T231323
Change-Id: If5282d8130b7aadb9c0861ec75eff4976cec7824