filebackend: Convert trigger_error to PSR log warning
* Add 'filebackend' name to the context.
* Convert message to a normalized form for better de-duplication
  in Logstash so that it is easier to see whether something is rare
  or common. Without this, each instance is unique and it might seem
  rare in a small sample of log messages.

This does not fully solve T231107, because we'd really want to
know at least the file name the issue is about, but it seems
the FileBackendStore does not have access to that right now,
that's for a future commit. (It also can't be logged from the
caller right now because this method does not expose the result
of the operation in question).

Bug: T231107
Change-Id: Iaf18d7d5c11334c84077667aa147d99fadb76a30