pipeline: rename publish stage to dev; add dev tag
Renamed `dev` stage of the `publish` pipeline to `publish` here:

  https://gerrit.wikimedia.org/r/c/mediawiki/core/+/539589

This turns out to have been unnecessary, and is confusing since the name
of the stage winds up as part of the default tag on the published
images, so reverting.

This patch also adds a "dev" tag to the image, so that we have a fixed
name for use by development environments.

Bug: T234379
Change-Id: I280ec935d5159f2992a62a5e7e33e71b6cbcd8a2