mediawiki.base: Remove mw.html.Cdata support
This was introduced in 2010 with e81c822c8b7e for use
by mw.loader to create <style> element, but has not been used
for that since 2012 with 57dda2b95e099.

It has never been used elsewhere to my knowledge.
In Codesearch, mwgrep, and mwgrep--user there are zero references
to "html.Cdata".

Bug: T233676
Change-Id: Ie99b41b6ee50cbc75e9258ef60ccae85e2cccf12