resourceloader: Remove listing of "load.php" from Special:Version
Doesn't seem particularly useful. Unlike api.php, this isn't
a public interface that you'd need the path of to do something.

It's path is configurable via wgLoadScript, but anything that needs
it already has access to it. It's mainly for internal purposes, like
thumb.php or opensearch_desc.php, which we don't list, either.

Change-Id: I52efbdc233451d0e7fd520f37b4dd642b104cc72