MessagesCu: Fix invalid characters in namespace aliases
In commit 5940aa6344693751345c8c062d1942e842a6e3fe in 2008 (SVN r35745),
these were accidentally changed from U+F011 (a private use area character)
to U+0011 (a control character). The control characters are certainly
incorrect, as they can't even be used in MediaWiki page titles. Having
a private use area character here is surprising to me, but it was
apparently used since 2006 as a substitute for U+A657, which was only
introduced in Unicode 5.1 in April 2008, hence the aliases.

I noticed this problem when I was doing some work analyzing MediaWiki's
date formats, and my editor warned me that this is a "binary file" :)

Change-Id: I5df29b0072d6dec8dcb8c7492d8b9623c93fdbf3