jquery.accessKeyLabel: Remove unused code for unsupported browsers
This was still providing support for various browsers where JS
is unconditionally disabled in the Startup module, hence the code
is never used.

Keep the information about them as historical note for now since
this file has become a historical record of sorts about
accessKeyLabel, which may be useful at some point.

* Remove old Opera code.
* Remove old Firefox code.
* Remove old Konqueror code.
* Remove old Safari code.
* Remove Safari on Windows code (macOS-only as of Safari 6).

Also:

* Rename internal name for getAccessKeyModifiers() parameter
  from 'ua' to 'nav' because it takes a Navigator object,
  not a userAgent string (per $.client.profile).

* Add test case for newer Konqueror.
  Copied from <https://github.com/wikimedia/jquery-client/commit/ae3e7fb294>.

Bug: T233676
Change-Id: I0cdbc3207f33e5b34f1d814c0b9b50f664b5b67c