RequestContext: Clarify uselang support and how it works
Follows-up to cad882954a08. I considered augmenting
the sanitizeLangCode() or isValidCode() function to reject
invalid UTF-8. But I realise now that it isn't so much about
"invalid" UTF-8 as it is about non-canonical forms for expressing
otherwise valid UTF-8 strings.

For example "Le\xcc\x81cole" and "L\xc3\xa9cole" both render
as "Lécole" but use different ways to accomplish it. To prevent
creation of inaccessible page titles, and to make link creation
easy, UtfNormal is used to normalises these when opening redlink
urls and when saving wikitext.

While non-ASCII will continue to work after cad882954a08, any
links to such characters must be made in normal form. Given
these are generally linked to from wiki pages or gadgets with
a predefined set of urls (and not typed ad-hoc by an end user)
this seems fine to require. I fid not find violations of this
on Commons.

Document the ceveat for now.

Change-Id: I36fece34e40f7e58b7aed5521c49a241a67d006f