Make partial block error message more generic
The partial block error message can be displayed while the blocked
user is attempting many different actions, so do not assume that they
are trying to edit. The user may be have a partial block that blocks
them from editing all articles, so do not state that they can
definitely edit other pages.

Bug: T233206
Change-Id: I31f643dd4a3f0f047e82466bc67082b8c1fb9475