resources: Allow loading Moment.js like a module
First of all, remove the hack with moment-global.js, it does nothing.
It was originally added because the Moment init code used to treat the
ResourceLoader environment like a Node.js environment, and did not
define a global `moment` variable. It is not needed since f13a0d9521,
where the Moment code changed so that it treats us like a normal
browser environment and defines the global.

Instead, add a hack to do the opposite, and allow loading Moment like
a module, using `require( 'moment' )`.

Change-Id: I9a59a6ad01850b30327e4215f2be61b8d1c41277